Split large frontend route files
