+++ frontend/src/pages/Transactions.tsx
@@ -2,807 +2,25 @@ import {
   useBanks,
   useBulkIgnoreMerchants,
   useBulkIgnoreTransactions,
-  useBuckets,
-  useCategorizeMerchant,
-  useCategories,
   useFacets,
-  useIgnoreByMerchant,
-  useIgnoreTransaction,
   useTransactions,
-  useUpdateTransactionDescription,
-  useUpdateTransactionFields,
 } from '@/api/queries'
-import type { BankColor, Transaction, TransactionFilters } from '@/api/types'
-import { DateRangePicker } from '@/components/DateRangePicker'
-import { FilterCombobox } from '@/components/FilterCombobox'
-import { InlineSelect } from '@/components/InlineSelect'
-import { LabelCombobox } from '@/components/LabelCombobox'
-import { LabelSearch } from '@/components/LabelSearch'
+import type { TransactionFilters } from '@/api/types'
 import { Pagination } from '@/components/Pagination'
-import { SlideNotification } from '@/components/SlideNotification'
 import { useDisplay } from '@/contexts/DisplayContext'
-import { cn, formatCurrency, formatDate, getSourceColor } from '@/lib/utils'
+import { useTooltip } from '@/hooks/useTooltip'
 import { formatNumberForLocale } from '@/i18n/format'
 import { useI18n } from '@/i18n/I18nProvider'
-import { ChevronDown, ChevronUp, EyeOff, Eye, Funnel, Plus, X } from 'lucide-react'
+import { cn, formatCurrency } from '@/lib/utils'
+import { ChevronDown, ChevronUp, Eye, EyeOff, Funnel, X } from 'lucide-react'
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
-import { createPortal } from 'react-dom'
 import { useSearchParams } from 'react-router-dom'
-import { useQueryClient } from '@tanstack/react-query'
-import { useTooltip } from '@/hooks/useTooltip'
-
-function useDebounce<T>(value: T, delay: number): T {
-  const [debounced, setDebounced] = useState(value)
-  useEffect(() => {
-    const timer = setTimeout(() => setDebounced(value), delay)
-    return () => clearTimeout(timer)
-  }, [value, delay])
-  return debounced
-}
-
-function AmountCell({ tx }: { tx: Transaction }) {
-  const { locale } = useI18n()
-  const hasOriginal =
-    tx.original_amount !== undefined &&
-    tx.original_currency !== undefined &&
-    tx.original_currency !== tx.currency
-
-  return (
-    <div className="text-right">
-      <div className="font-mono text-sm tabular-nums text-primary">
-        {formatCurrency(tx.amount, tx.currency, locale)}
-      </div>
-      {hasOriginal && (
-        <div className="font-mono text-[10px] tabular-nums text-muted-foreground">
-          {formatCurrency(tx.original_amount!, tx.original_currency!, locale)}
-          {tx.exchange_rate !== undefined && ` @ ${tx.exchange_rate.toFixed(2)}`}
-        </div>
-      )}
-    </div>
-  )
-}
-
-function HighlightedText({ text, query }: { text: string; query: string }) {
-  const normalizedQuery = query.trim()
-  if (!normalizedQuery) return <>{text}</>
-
-  const index = text.toLowerCase().indexOf(normalizedQuery.toLowerCase())
-  if (index === -1) return <>{text}</>
-
-  return (
-    <>
-      {text.slice(0, index)}
-      <mark className="rounded bg-primary/20 px-0.5 text-primary">
-        {text.slice(index, index + normalizedQuery.length)}
-      </mark>
-      {text.slice(index + normalizedQuery.length)}
-    </>
-  )
-}
-
-function DescriptionCell({ tx, searchQuery }: { tx: Transaction; searchQuery: string }) {
-  const [editing, setEditing] = useState(false)
-  const [value, setValue] = useState(tx.description)
-  const inputRef = useRef<HTMLInputElement>(null)
-  const { mutate: updateDesc } = useUpdateTransactionDescription()
-  const { handlers: descTip, tip: descTipEl } = useTooltip()
-
-  useEffect(() => {
-    setValue(tx.description)
-  }, [tx.description])
-
-  useEffect(() => {
-    if (editing) inputRef.current?.focus()
-  }, [editing])
-
-  const commit = useCallback(() => {
-    if (value !== tx.description) {
-      updateDesc({ id: tx.id, description: value })
-    }
-    setEditing(false)
-  }, [value, tx.description, tx.id, updateDesc])
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') commit()
-    if (e.key === 'Escape') {
-      setValue(tx.description)
-      setEditing(false)
-    }
-  }
-
-  if (editing) {
-    return (
-      <input
-        ref={inputRef}
-        type="text"
-        value={value}
-        onChange={(e) => setValue(e.target.value)}
-        onBlur={commit}
-        onKeyDown={handleKeyDown}
-        className="w-full rounded-sm border border-primary bg-accent px-2 py-1 text-xs text-foreground focus:outline-none focus:ring-1 focus:ring-ring"
-        aria-label="Edit description"
-      />
-    )
-  }
-
-  return (
-    <>
-      <button
-        onClick={() => setEditing(true)}
-        {...descTip(value || 'Click to add description')}
-        className="w-full truncate text-left text-xs text-muted-foreground transition-colors hover:text-foreground"
-        aria-label={`Edit description: ${value || 'empty'}`}
-      >
-        {value ? (
-          <HighlightedText text={value} query={searchQuery} />
-        ) : (
-          <span className="opacity-30">—</span>
-        )}
-      </button>
-      {descTipEl}
-    </>
-  )
-}
-
-function LabelsCell({ tx }: { tx: Transaction }) {
-  return <LabelCombobox tx={tx} />
-}
-
-type IgnoreReasonAction = {
-  label: string
-  variant?: 'primary' | 'secondary'
-  onClick: () => void
-  disabled?: boolean
-}
-
-function IgnoreReasonModal({
-  title,
-  subtitle,
-  context,
-  placeholder,
-  reason,
-  onReasonChange,
-  actions,
-  onCancel,
-}: {
-  title: string
-  subtitle?: string
-  context?: string
-  placeholder?: string
-  reason: string
-  onReasonChange: (value: string) => void
-  actions: IgnoreReasonAction[]
-  onCancel: () => void
-}) {
-  const { t } = useI18n()
-  const inputPlaceholder = placeholder ?? t('common.reasonOptional')
-
-  return createPortal(
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-background/80 backdrop-blur-sm"
-      onMouseDown={(event) => {
-        if (event.target === event.currentTarget) onCancel()
-      }}
-    >
-      <div className="w-full max-w-sm space-y-4 rounded-lg border border-border bg-card p-6 shadow-xl">
-        <h2 className="text-sm font-semibold text-foreground">{title}</h2>
-        {subtitle && <p className="text-xs text-muted-foreground">{subtitle}</p>}
-        {context && <p className="text-xs text-muted-foreground">{context}</p>}
-        <input
-          autoFocus
-          value={reason}
-          onChange={(event) => onReasonChange(event.target.value)}
-          onKeyDown={(event) => {
-            if (event.key === 'Escape') onCancel()
-          }}
-          placeholder={inputPlaceholder}
-          className="w-full rounded border border-border bg-secondary px-3 py-2 text-sm text-foreground placeholder:text-muted-foreground focus:border-primary focus:outline-none focus:ring-1 focus:ring-ring"
-        />
-        <div className="flex flex-col gap-2">
-          {actions.map((action) => (
-            <button
-              key={action.label}
-              onClick={action.onClick}
-              disabled={action.disabled}
-              className={cn(
-                'w-full rounded-md px-4 py-2 text-sm transition-colors',
-                action.variant === 'primary'
-                  ? 'bg-primary text-primary-foreground hover:bg-primary/90'
-                  : 'border border-border text-muted-foreground hover:text-foreground',
-                action.disabled && 'cursor-not-allowed opacity-50',
-              )}
-            >
-              {action.label}
-            </button>
-          ))}
-          <button
-            onClick={onCancel}
-            className="w-full px-4 py-1 text-xs text-muted-foreground hover:text-foreground"
-          >
-            {t('common.cancel')}
-          </button>
-        </div>
-      </div>
-    </div>,
-    document.body,
-  )
-}
-
-function CategoryBucketCell({ tx }: { tx: Transaction }) {
-  const { data: categories = [] } = useCategories()
-  const { data: buckets = [] } = useBuckets()
-  const { mutate: updateFields } = useUpdateTransactionFields()
-  const { mutate: categorizeMerchant } = useCategorizeMerchant()
-  const [pendingMerchantCat, setPendingMerchantCat] = useState<{
-    merchant: string
-    category: string
-    bucket: string
-  } | null>(null)
-  const [applyError, setApplyError] = useState(false)
-
-  return (
-    <div className="flex flex-col gap-0.5">
-      <InlineSelect
-        value={tx.category}
-        options={categories.map((c) => c.name)}
-        onCommit={(category) =>
-          updateFields(
-            { id: tx.id, patch: { category } },
-            {
-              onSuccess: () => {
-                if (tx.merchant_info) {
-                  setPendingMerchantCat({ merchant: tx.merchant_info, category, bucket: tx.bucket })
-                }
-              },
-            },
-          )
-        }
-        className="text-foreground"
-      />
-      <InlineSelect
-        value={tx.bucket}
-        options={buckets.map((b) => b.name)}
-        onCommit={(bucket) =>
-          updateFields(
-            { id: tx.id, patch: { bucket } },
-            {
-              onSuccess: () => {
-                if (tx.merchant_info) {
-                  setPendingMerchantCat({
-                    merchant: tx.merchant_info,
-                    category: tx.category,
-                    bucket,
-                  })
-                }
-              },
-            },
-          )
-        }
-        className="text-muted-foreground"
-      />
-
-      {pendingMerchantCat && (
-        <SlideNotification
-          key={`${pendingMerchantCat.merchant}:${pendingMerchantCat.category}:${pendingMerchantCat.bucket}`}
-          actions={[
-            { label: 'Dismiss', value: false },
-            { label: 'Apply', value: true, primary: true },
-          ]}
-          defaultActionIndex={0}
-          onAction={(apply) => {
-            if (apply && pendingMerchantCat) {
-              categorizeMerchant(
-                {
-                  merchant: pendingMerchantCat.merchant,
-                  category: pendingMerchantCat.category,
-                  bucket: pendingMerchantCat.bucket,
-                },
-                {
-                  onSuccess: () => setPendingMerchantCat(null),
-                  onError: () => {
-                    setPendingMerchantCat(null)
-                    setApplyError(true)
-                  },
-                },
-              )
-            } else {
-              setPendingMerchantCat(null)
-            }
-          }}
-        >
-          Apply{' '}
-          <span className="font-medium text-foreground">
-            &quot;{pendingMerchantCat.category || '(none)'}&quot;
-          </span>
-          {' / '}
-          <span className="font-medium text-foreground">
-            {pendingMerchantCat.bucket || '(none)'}
-          </span>{' '}
-          to all <span className="font-medium text-foreground">{pendingMerchantCat.merchant}</span>{' '}
-          transactions?
-        </SlideNotification>
-      )}
-
-      {applyError && (
-        <SlideNotification
-          key="categorize-error"
-          actions={[{ label: 'Dismiss', value: false }]}
-          defaultActionIndex={0}
-          onAction={() => setApplyError(false)}
-        >
-          Failed to apply category to all transactions. Please try again.
-        </SlideNotification>
-      )}
-    </div>
-  )
-}
-
-function SourceCell({ tx, banks }: { tx: Transaction; banks?: BankColor[] }) {
-  if (!tx.source.bank && !tx.source.type) {
-    return <span className="text-xs text-muted-foreground">—</span>
-  }
-
-  const bankColor = getSourceColor(tx.source.bank, banks)
-
-  return (
-    <div className="flex min-w-0 flex-col gap-0.5 leading-tight">
-      {tx.source.bank && (
-        <span className="flex min-w-0 items-center gap-1.5">
-          <span
-            className="h-2 w-2 flex-shrink-0 rounded-full shadow-[0_0_0_2px_hsl(var(--background))]"
-            style={{ backgroundColor: bankColor }}
-            aria-hidden="true"
-          />
-          <span className="truncate font-medium text-foreground">{tx.source.bank}</span>
-        </span>
-      )}
-      {tx.source.type && (
-        <span
-          className={cn(
-            'block truncate text-[11px] text-muted-foreground',
-            tx.source.bank && 'pl-3.5',
-          )}
-        >
-          {tx.source.type}
-        </span>
-      )}
-    </div>
-  )
-}
-
-function IgnoreButton({ tx }: { tx: Transaction }) {
-  const { t } = useI18n()
-  const { mutate: ignoreTransaction, isPending } = useIgnoreTransaction()
-  const { mutate: ignoreByMerchant } = useIgnoreByMerchant()
-  const qc = useQueryClient()
-  const { handlers: ignoreTip, tip: ignoreTipEl } = useTooltip()
-  const [showModal, setShowModal] = useState(false)
-  const [reason, setReason] = useState('')
-
-  const handleIconClick = () => {
-    if (tx.muted) {
-      ignoreTransaction(
-        { id: tx.id, muted: false },
-        { onSuccess: () => qc.invalidateQueries({ queryKey: ['transactions'] }) },
-      )
-      return
-    }
-    setReason('')
-    setShowModal(true)
-  }
-
-  const handleJustThis = () => {
-    setShowModal(false)
-    ignoreTransaction(
-      { id: tx.id, muted: true, reason },
-      { onSuccess: () => qc.invalidateQueries({ queryKey: ['transactions'] }) },
-    )
-  }
-
-  const handleIgnoreMerchant = () => {
-    setShowModal(false)
-    // Also ignore this individual transaction.
-    ignoreTransaction({ id: tx.id, muted: true, reason })
-    // Ignore all same-merchant transactions.
-    ignoreByMerchant({ pattern: tx.merchant_info, reason })
-  }
-
-  const label = tx.muted
-    ? t('transactions.ignore.tooltip.restore')
-    : t('transactions.ignore.tooltip.ignore')
-
-  return (
-    <>
-      <button
-        onClick={handleIconClick}
-        disabled={isPending}
-        aria-label={label}
-        {...(!showModal ? ignoreTip(label) : {})}
-        className={cn(
-          'flex items-center rounded p-1 transition-colors',
-          tx.muted
-            ? 'text-warning hover:text-foreground'
-            : 'text-muted-foreground/30 hover:text-muted-foreground',
-        )}
-      >
-        {tx.muted ? <EyeOff size={13} /> : <Eye size={13} />}
-      </button>
-      {!showModal && ignoreTipEl}
-
-      {showModal && (
-        <IgnoreReasonModal
-          title={t('transactions.ignore.transactionTitle')}
-          subtitle={t('transactions.ignore.transactionDescription')}
-          context={tx.merchant_info ? `${t('common.merchant')}: ${tx.merchant_info}` : undefined}
-          reason={reason}
-          onReasonChange={setReason}
-          onCancel={() => setShowModal(false)}
-          actions={[
-            ...(tx.merchant_info
-              ? [
-                  {
-                    label: t('transactions.ignore.merchantAction'),
-                    variant: 'primary' as const,
-                    onClick: handleIgnoreMerchant,
-                  },
-                ]
-              : []),
-            {
-              label: t('transactions.ignore.thisTransaction'),
-              onClick: handleJustThis,
-            },
-          ]}
-        />
-      )}
-    </>
-  )
-}
-
-function TransactionRow({
-  tx,
-  banks,
-  selected,
-  onToggleSelect,
-  searchQuery,
-}: {
-  tx: Transaction
-  banks?: BankColor[]
-  selected: boolean
-  onToggleSelect: () => void
-  searchQuery: string
-}) {
-  const { timezone, timeFormat } = useDisplay()
-  const { locale } = useI18n()
-  return (
-    <tr
-      className={cn(
-        'border-b border-border transition-colors hover:bg-accent/50',
-        selected && 'bg-secondary/30',
-        tx.muted && 'opacity-40',
-      )}
-    >
-      <td className="px-0 py-2.5 text-center">
-        <input
-          type="checkbox"
-          checked={selected}
-          onChange={onToggleSelect}
-          aria-label={`Select transaction ${tx.merchant_info || tx.message_id}`}
-        />
-      </td>
-      <td className="overflow-hidden whitespace-nowrap px-3 py-2.5">
-        <span className="block truncate font-mono text-xs text-muted-foreground">
-          {formatDate(tx.timestamp, true, timezone, timeFormat, locale)}
-        </span>
-      </td>
-      <td className="overflow-hidden px-3 py-2.5">
-        <span className="block truncate text-sm text-foreground">
-          <HighlightedText text={tx.merchant_info} query={searchQuery} />
-        </span>
-      </td>
-      <td className="overflow-hidden whitespace-nowrap px-3 py-2.5">
-        <SourceCell tx={tx} banks={banks} />
-      </td>
-      <td className="px-3 py-2.5">
-        <AmountCell tx={tx} />
-      </td>
-      <td className="overflow-hidden px-3 py-2.5">
-        <CategoryBucketCell tx={tx} />
-      </td>
-      <td className="overflow-hidden px-3 py-2.5">
-        <LabelsCell tx={tx} />
-      </td>
-      <td className="overflow-hidden px-3 py-2.5">
-        <DescriptionCell tx={tx} searchQuery={searchQuery} />
-      </td>
-      <td className="sticky right-0 bg-card px-3 py-2.5 shadow-[-8px_0_12px_-12px_rgba(0,0,0,0.6)]">
-        <IgnoreButton tx={tx} />
-      </td>
-    </tr>
-  )
-}
-
-// ─── Filter panel with default + "Add filter" extras ────────────────────────
-
-type FilterPanelProps = {
-  filters: TransactionFilters
-  facets:
-    | {
-        sources: string[]
-        source_types: string[]
-        banks: string[]
-        categories: string[]
-        currencies: string[]
-        labels: string[]
-        buckets: string[]
-      }
-    | undefined
-  hourFrom: number | undefined
-  hourTo: number | undefined
-  updateFilter: (key: keyof Omit<TransactionFilters, 'page' | 'page_size'>, value: string) => void
-  updateParams: (updates: Record<string, string | undefined>) => void
-}
-
-const EXTRA_FILTER_KEYS = ['category', 'bucket', 'currency', 'hour_from', 'hour_to'] as const
-type ExtraFilterKey = (typeof EXTRA_FILTER_KEYS)[number]
-
-const EXTRA_FILTER_LABELS: Record<ExtraFilterKey, string> = {
-  category: 'Category',
-  bucket: 'Bucket',
-  currency: 'Currency',
-  hour_from: 'Hour from',
-  hour_to: 'Hour to',
-}
-
-function FilterPanel({
-  filters,
-  facets,
-  hourFrom,
-  hourTo,
-  updateFilter,
-  updateParams,
-}: FilterPanelProps) {
-  const { t } = useI18n()
-  const [addOpen, setAddOpen] = useState(false)
-  const [addHighlighted, setAddHighlighted] = useState(-1)
-  const addBtnRef = useRef<HTMLButtonElement>(null)
-  const [addPos, setAddPos] = useState<{ x: number; y: number } | null>(null)
-  const { handlers: addTip, tip: addTipEl } = useTooltip()
-
-  // Which extra filters are visible — an extra filter is visible if it has an active value,
-  // or if the user has explicitly added it via "Add filter".
-  const [addedFilters, setAddedFilters] = useState<Set<ExtraFilterKey>>(() => {
-    const pre = new Set<ExtraFilterKey>()
-    if (filters.category) pre.add('category')
-    if (filters.bucket) pre.add('bucket')
-    if (filters.currency) pre.add('currency')
-    if (hourFrom !== undefined) pre.add('hour_from')
-    if (hourTo !== undefined) pre.add('hour_to')
-    return pre
-  })
-
-  const visibleExtras = EXTRA_FILTER_KEYS.filter(
-    (k) =>
-      addedFilters.has(k) ||
-      (k === 'category' && filters.category) ||
-      (k === 'bucket' && filters.bucket) ||
-      (k === 'currency' && filters.currency) ||
-      (k === 'hour_from' && hourFrom !== undefined) ||
-      (k === 'hour_to' && hourTo !== undefined),
-  )
-
-  const hiddenExtras = EXTRA_FILTER_KEYS.filter((k) => !visibleExtras.includes(k))
-
-  const openAdd = () => {
-    const rect = addBtnRef.current?.getBoundingClientRect()
-    if (rect) setAddPos({ x: rect.left, y: rect.bottom + 4 })
-    setAddOpen(true)
-    setAddHighlighted(-1)
-  }
-
-  const closeAdd = () => {
-    setAddOpen(false)
-    setAddHighlighted(-1)
-  }
-
-  const selectExtraFilter = (key: ExtraFilterKey) => {
-    setAddedFilters((s) => new Set([...s, key]))
-    closeAdd()
-  }
-
-  useEffect(() => {
-    if (!addOpen) return
-    const handler = (e: MouseEvent) => {
-      const portal = document.getElementById('add-filter-portal')
-      if (!portal?.contains(e.target as Node) && !addBtnRef.current?.contains(e.target as Node)) {
-        closeAdd()
-      }
-    }
-    document.addEventListener('mousedown', handler)
-    return () => document.removeEventListener('mousedown', handler)
-  }, [addOpen])
-
-  return (
-    <div className="flex flex-wrap items-center gap-2 rounded-lg border border-border bg-card p-3">
-      {/* Always-visible: Date */}
-      <DateRangePicker
-        value={{
-          from: filters.date_from ? new Date(filters.date_from) : undefined,
-          to: filters.date_to ? new Date(filters.date_to) : undefined,
-        }}
-        onChange={(range) =>
-          updateParams({
-            date_from: range.from ? range.from.toISOString() : undefined,
-            date_to: range.to ? range.to.toISOString() : undefined,
-            page: '1',
-          })
-        }
-      />
-
-      {/* Always-visible: Source type */}
-      <FilterCombobox
-        value={filters.source_type ?? ''}
-        onChange={(v) => updateFilter('source_type', v)}
-        options={facets?.source_types ?? []}
-        placeholder={t('common.type')}
-        label={t('transactions.filter.sourceType')}
-      />
-
-      {/* Always-visible: Bank */}
-      <FilterCombobox
-        value={filters.bank ?? ''}
-        onChange={(v) => updateFilter('bank', v)}
-        options={facets?.banks ?? []}
-        placeholder={t('common.bank')}
-        label={t('transactions.filter.bank')}
-      />
-
-      {/* Always-visible: Label */}
-      <LabelSearch
-        value={filters.label ?? ''}
-        onChange={(v) => updateFilter('label', v)}
-        options={facets?.labels ?? []}
-      />
-
-      {/* Extra filters added by user */}
-      {visibleExtras.includes('category') && (
-        <FilterCombobox
-          value={filters.category ?? ''}
-          onChange={(v) => updateFilter('category', v)}
-          options={facets?.categories ?? []}
-          placeholder="Category"
-          label="Filter by category"
-        />
-      )}
-      {visibleExtras.includes('bucket') && (
-        <FilterCombobox
-          value={filters.bucket ?? ''}
-          onChange={(v) => updateFilter('bucket', v)}
-          options={facets?.buckets ?? []}
-          placeholder="Bucket"
-          label="Filter by bucket"
-        />
-      )}
-      {visibleExtras.includes('currency') && (
-        <FilterCombobox
-          value={filters.currency ?? ''}
-          onChange={(v) => updateFilter('currency', v)}
-          options={facets?.currencies ?? []}
-          placeholder="Currency"
-          label="Filter by currency"
-        />
-      )}
-      {visibleExtras.includes('hour_from') && (
-        <FilterCombobox
-          value={hourFrom !== undefined ? `${String(hourFrom).padStart(2, '0')}:00` : ''}
-          onChange={(v) =>
-            updateParams({
-              hour_from: v ? String(parseInt(v.split(':')[0], 10)) : undefined,
-              page: '1',
-            })
-          }
-          options={Array.from({ length: 24 }, (_, i) => `${String(i).padStart(2, '0')}:00`)}
-          placeholder="Hour from"
-          label="Filter from hour (DB timezone)"
-        />
-      )}
-      {visibleExtras.includes('hour_to') && (
-        <FilterCombobox
-          value={hourTo !== undefined ? `${String(hourTo).padStart(2, '0')}:59` : ''}
-          onChange={(v) =>
-            updateParams({
-              hour_to: v ? String(parseInt(v.split(':')[0], 10)) : undefined,
-              page: '1',
-            })
-          }
-          options={Array.from({ length: 24 }, (_, i) => `${String(i).padStart(2, '0')}:59`)}
-          placeholder="Hour to"
-          label="Filter to hour (DB timezone)"
-        />
-      )}
-
-      {/* Add filter button */}
-      {hiddenExtras.length > 0 && (
-        <>
-          <button
-            ref={addBtnRef}
-            onClick={openAdd}
-            onKeyDown={(event) => {
-              if (event.key === 'Escape') {
-                event.preventDefault()
-                closeAdd()
-                return
-              }
-              if (event.key === 'ArrowDown') {
-                event.preventDefault()
-                if (!addOpen) {
-                  openAdd()
-                  setAddHighlighted(0)
-                  return
-                }
-                setAddHighlighted((current) => Math.min(current + 1, hiddenExtras.length - 1))
-                return
-              }
-              if (event.key === 'ArrowUp') {
-                event.preventDefault()
-                if (!addOpen) {
-                  openAdd()
-                  setAddHighlighted(hiddenExtras.length - 1)
-                  return
-                }
-                setAddHighlighted((current) => Math.max(current - 1, 0))
-                return
-              }
-              if (event.key === 'Enter' && addOpen && addHighlighted >= 0) {
-                event.preventDefault()
-                selectExtraFilter(hiddenExtras[addHighlighted])
-              }
-            }}
-            aria-haspopup="menu"
-            aria-expanded={addOpen}
-            {...addTip('Add filter')}
-            className="flex items-center gap-1 rounded-md border border-dashed border-border px-2 py-1.5 text-xs text-muted-foreground transition-colors hover:border-primary hover:text-primary"
-          >
-            <Plus size={11} />
-            Add filter
-          </button>
-          {addTipEl}
-        </>
-      )}
-
-      {/* Add-filter dropdown portal */}
-      {addOpen &&
-        addPos &&
-        createPortal(
-          <div
-            id="add-filter-portal"
-            role="menu"
-            aria-label="Add filter options"
-            className="fixed z-50 min-w-[160px] rounded-lg border border-border bg-card py-1 shadow-xl"
-            style={{ left: addPos.x, top: addPos.y }}
-          >
-            {hiddenExtras.map((key, index) => (
-              <button
-                key={key}
-                type="button"
-                role="menuitem"
-                onClick={() => {
-                  selectExtraFilter(key)
-                }}
-                onMouseEnter={() => setAddHighlighted(index)}
-                className={cn(
-                  'flex w-full items-center gap-2 px-3 py-2 text-xs text-muted-foreground hover:bg-accent hover:text-foreground',
-                  index === addHighlighted && 'bg-accent text-foreground',
-                )}
-              >
-                {EXTRA_FILTER_LABELS[key]}
-              </button>
-            ))}
-          </div>,
-          document.body,
-        )}
-    </div>
-  )
-}
+import {
+  FilterPanel,
+  IgnoreReasonModal,
+  TransactionRow,
+  useDebounce,
+} from './transactions/TransactionsParts'
 
 export function Transactions() {
   const [searchParams, setSearchParams] = useSearchParams()
